Restructured slurm collector
